% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{ribbon}
\alias{ribbon}
\title{Typewriter Ribbon Data}
\format{A data frame with columns:
\itemize{
  \item \code{V1} absolute frequency reflecting strong preference for corresponding \code{obj1}
  \item \code{V2} absolute frequency reflecting moderate preference for corresponding \code{obj1}
  \item \code{V3} absolute frequency reflecting weak preference for corresponding \code{obj1}
  \item \code{V4} absolute frequency reflecting no preference
  \item \code{V5} absolute frequency reflecting weak preference for corresponding \code{obj2}
  \item \code{V6} absolute frequency reflecting moderate preference for corresponding \code{obj2}
  \item \code{V7} absolute frequency reflecting strong preference for corresponding \code{obj2}
  \item \code{obj1} integer 1, ..., 5 referring to the 5 ribbon brands
  \item \code{obj2} integer 1, ..., 5 referring to the 5 ribbon brands
}}
\source{
Agresti A (1992). "Analysis of ordinal paired comparison data."
Applied Statistics, pp. 287-297. Table 1.
}
\usage{
data(ribbon)
}
\description{
The typwriter ribbon data used in the paper by Agresti (1992), where 5
different typewriter ribbon brands were compared pairwise by 30 secretaries.
}
\references{
Agresti A (1992). "Analysis of ordinal paired comparison data."
Applied Statistics, pp. 287-297.
}
\keyword{datasets}

